remounts pageInput from within pageInputWrapper successfully!
