+++ src/components/Dashboard/Input/PageInputWrapper.js
@@ -1,49 +1,24 @@
 import './_PageInput.css'
 import React, { useState, useEffect, useRef } from 'react'
+import PageInput from './PageInput'
 // compos
 // import PageInput from './PageInput'
 
 function PageInputWrapper({ pushPage, nestSeq }) {
 
-    const [page, setPage] = useState('')
     const [addingPage, setAddingPage] = useState(false)
-    const [lock, setLock] = useState(false)
 
-    const pageInputRef = useRef()
-
-    useEffect(() => {
-        if(addingPage) {
-            setLock(false)
-            setTimeout(() => {
-                pageInputRef.current.focus()
-            }, .0001)
-        }
-    }, [addingPage])
-
-    const handleWrite = (ev) => {
-        setPage(ev.target.value)
-    }
-
-    const handleBlur = () => {
-        if(!lock) {
-            setAddingPage(false)
-        }
-    }
-
-    const lockUnlock = () => {
-        setLock(true)
-        setTimeout(() => {
-            setLock(false)
-        }, .0001)
-    }
-
-    const handleSave = () => {
+    const handleSave = (nestSeq) => (page) => {
 
         if (page.length > 0) {
             pushPage(nestSeq ? nestSeq[nestSeq.length-1] : null, page)
             setAddingPage(false)
         }
 
+    }
+
+    const unMountOnBlur = () => {
+        setAddingPage(false)
     }
 
 return (
@@ -60,26 +35,11 @@ return (
     )
     :
     (
-        // this part needs to be it's own compo
-        // poma 2
-        // 1.19.21
 
-        <div 
-        className='pageInput_addingPage'
-        onMouseDown={(ev) => ev.preventDefault()}
-        onBlur={handleBlur}
-        onClick={lockUnlock}
-        >
-            <input 
-            className='addingPage_input'
-            onChange={handleWrite}
-            value={page}
-            ref={pageInputRef}
-            />
-            <div 
-            className='addingPage_saveBtn'
-            onClick={handleSave}>+ Save Page</div>
-        </div>
+        <PageInput
+        unMountOnBlur={unMountOnBlur}
+        handleSave={handleSave(nestSeq)}
+        />
     )
 }
 </>
